refactor: move search bar to search dir
